Sort project tree file
