+++ Library/ChiriUtils/Utils/Extensions/Logic/Array.swift
@@ -1,0 +1,9 @@
+//
+//  Array.swift
+//  ChiriUtils
+//
+//  Created by Mauricio Chirino on 1/4/20.
+//  Copyright © 2020 Mauricio Chirino. All rights reserved.
+//
+
+import Foundation
